updates delays for video and audio recorders

+++ realTimeModules/videoRecorder/videoRecorder.py
@@ -39,7 +39,7 @@ def localVideoRecorder(frameQ, videoInputSource):
 
         cv2.putText(frame, "frame : "+str(frameCounter), (30,30), cv2.FONT_HERSHEY_PLAIN, 1.5, 255)
 
-        time.sleep(1/float(35))
+        time.sleep(1/float(50))
 
         # display frame
         cv2.imshow("feed", frame)
